sam-deploy/finops-backend/api/requirements.txt
fastapi==0.115.5
mangum==0.17.0

# Pydantic v2 stack
pydantic==2.9.2
pydantic-core==2.23.4
typing-extensions==4.8.0

# Local server (harmless on Lambda)
uvicorn[standard]==0.22.0

# AWS + HTTP
boto3>=1.34,<2
# botocore: let boto3 pull the compatible version
requests==2.31.0
httpx==0.24.1
python-multipart>=0.0.9
python-dotenv>=1.0.1

# Needed by your code
SQLAlchemy>=2.0.30
reportlab>=3.6.12
python-jose[cryptography]>=3.3.0
